# API/gin_api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hoseo NFC 2021 projct API Page",
        "title": "NFC API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/v1/common/visitHistory/{Wearable_SN}": {
            "get": {
                "description": "테스트용 작성후 삭제예정",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "app working test",
                "operationId": "Wearable_SN",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Wearable_SN",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.UserLogModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.UserLogModel": {
            "type": "object",
            "properties": {
                "Date": {
                    "type": "string",
                    "example": "2021-06-30"
                },
                "IDX": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "Time": {
                    "type": "string",
                    "example": "15:30:30"
                },
                "building_name": {
                    "type": "string"
                },
                "temp": {
                    "type": "string",
                    "example": "36.5"
                }
            }
        }
    }
}